Do not reimplement x11rb utilities
The structs FreePixmap and FreeGC contain a reference to a pixmap or a
graphics context that is freed on drop. Such utilities are already
available from x11rb under the names PixmapWrapper and GcontextWrapper.

This commit switches the code to use the version from x11rb.

No "real" behavioural changes intended. However:

- The old code first created a FreePixmap instance and then called
  create_pixmap(). If create_pixmap() failed, the FreePixmap request
  would result in an X11 error since it tried to free a non-existing
  pixmap.

- The old code immediately called check() on the cookies for the
  CreatePixmap request. This commit first creates both pixmaps and then
  checks for errors, thus saving one round-trip to the X11 server.

Signed-off-by: Uli Schlachter <psychon@znc.in>