Simplify impls of PartialEq, Eq, ParialOrd for Area
The hand-written impls for PartialEq and Eq are equivalent to the
compiler-generated ones. The implementation of PartialOrd can just
forward to the implementation of Ord.

In theory, one could also derive PartialOrd and Ord, but that would
require reordering the fields of the struct to keep the current
behaviour. From [0]:

  When derived on structs, it will produce a lexicographic ordering
  based on the top-to-bottom declaration order of the struct’s members.

[0]: https://doc.rust-lang.org/std/cmp/trait.PartialOrd.html#derivable

Signed-off-by: Uli Schlachter <psychon@znc.in>