Overhaul code around GetProperty requests
The code for get_client_list() and get_window_state() was quite similar,
so I factored out a helper function get_property32() that gets the
contents of a property with format 32.

In the process, I also changed the code slightly. Instead of manually
decoding the value as little endian (which would be incorrect on big
endian hosts), this uses the value32() helper method on
GetPropertyReply. This returns an Option<Iterator<Item=u32>> containing
the individual pieces of the reply.

However, this also checks that the "format" field is set to 32 and
otherwise returns None. Properties of type WINDOW and ATOM should always
be in format 32. Formats 16 or 8 do not make sense here and were
interpreted incorrectly by the old code anyway. Thus, I think it makes
sense to add this new error case here.

I also touched the code in get_window_name(). The property _NET_WM_NAME
should always have type UTF8_STRING (which this code does not check, but
whatever...). However, this code treats malformed UTF8 as latin1 for
some reason.

Anyway, what I did to this function: x11rb provides the property's
contents as a Vec<u8>. The old code was using std::str::from_utf8() to
get a &str from this and then used String::into() to convert this into a
string. This always copied the bytes. By using String::from_utf8(), this
can directly re-use the allocation from the original Vec<u8> without
needing to copy. If this fails, decode_latin1() is still used as before.

Signed-off-by: Uli Schlachter <psychon@znc.in>