Remove "last_clipboard" variable
last_clipboard was initially set to None. The only place setting it to
Some was right before the check whether it contained a value.

This commit gets rid of the last_clipboard variable and instead merges
the push of the event to the event list with the code that creates the
event.

Signed-off-by: Uli Schlachter <psychon@znc.in>